move common used by trigger to core module
