+++ core/src/main/java/com/kii/thingif/core/trigger/TriggersWhen.java
@@ -1,4 +1,4 @@
-package com.kii.thingif.trigger;
+package com.kii.thingif.core.trigger;
 
 public enum TriggersWhen {
     CONDITION_TRUE,
